+++ src/collections/domain/repositories/ICollectionsRepository.ts
@@ -12,6 +12,7 @@ import { CollectionItemType } from '../../../collections/domain/models/Collectio
 import { CollectionLinks } from '../models/CollectionLinks'
 import { CollectionSummary } from '../models/CollectionSummary'
 import { LinkingObjectType } from '../useCases/GetCollectionsForLinking'
+import { StorageDriver } from '../../../datasets/domain/models/StorageDriver'
 
 export interface ICollectionsRepository {
   getCollection(collectionIdOrAlias: number | string): Promise<Collection>
@@ -68,4 +69,8 @@ export interface ICollectionsRepository {
     searchTerm: string,
     alreadyLinked: boolean
   ): Promise<CollectionSummary[]>
+  getCollectionStorageDriver(
+    collectionIdOrAlias: number | string,
+    getEffective?: boolean
+  ): Promise<StorageDriver>
 }